Added $lineheightmod to ktype, Added kunit function
Added $lineheightmod parameter to the ktype mixin, to replace push and pull in the future with a single value. It takes negative and positive integers like the scale unit, corresponding to half-line modificaitons to the lineheight of the output typographic element. eg. use ktype(0,1,0,2) to add a whole baseline unit to the lineheight.

Added kunit function for non-type vertical rythm arrangements, simply declare kunit(X) to output X baseline height in pixels, it also accepts a second [boolean] parameter to output in rem units instead. eg. kunit(2, true) for 2 baseline units equivalent in rems. This is useful for spacing and sizing your layout units and rapid prototyping of frameworks.