enabled lineheightmod for 0 scale ktype + fix for 'may only compare numbers' using libsass
defaulted push and pull to '0' rather than 'false' in ktype() mixin, in which @if $pull != 0 comparison is made later, which might be libsass. The bug might be unrelated but the commit is just logical anyhow, no need to false push/pull, just use the intended numerical.

Also enabled lineheightmod for 0 scale ktype, in case one wants double spaced body sized type for example.